backend-flask/requirements.txt
aniso8601==9.0.1
apispec==6.2.0
black==23.1.0
click==8.1.3
Flask==2.2.3
flask-apispec==0.11.4
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-WTF==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.0
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
webargs==8.2.0
Werkzeug==2.2.3
WTForms==3.0.1
